refactor: move `Client` impl blocks out of test helper modules [WPB-15530]
