+++ crypto/src/test_utils/context.rs
@@ -38,7 +38,6 @@ use core_crypto_keystore::connection::FetchFromDatabase;
 use core_crypto_keystore::entities::{
     EntityFindParams, MlsCredential, MlsEncryptionKeyPair, MlsHpkePrivateKey, MlsKeyPackage, MlsSignatureKeyPair,
 };
-use mls_crypto_provider::MlsCryptoProvider;
 use openmls::prelude::{
     Credential, CredentialWithKey, CryptoConfig, ExternalSender, HpkePublicKey, KeyPackage, KeyPackageIn,
     LeafNodeIndex, Lifetime, MlsMessageIn, QueuedProposal, SignaturePublicKey, StagedCommit,
@@ -824,43 +823,5 @@ impl MlsConversation {
 
     pub fn extensions(&self) -> &openmls::prelude::Extensions {
         self.group.export_group_context().extensions()
-    }
-}
-
-impl Client {
-    pub(crate) async fn init_x509_credential_bundle_if_missing(
-        &self,
-        backend: &MlsCryptoProvider,
-        sc: SignatureScheme,
-        cb: CertificateBundle,
-    ) -> Result<()> {
-        let existing_cb = self
-            .find_most_recent_credential_bundle(sc, MlsCredentialType::X509)
-            .await
-            .is_err();
-        if existing_cb {
-            self.save_new_x509_credential_bundle(&backend.keystore(), sc, cb)
-                .await
-                .map_err(RecursiveError::mls_client("saving new x509 credential bundle"))?;
-        }
-        Ok(())
-    }
-
-    pub(crate) async fn generate_one_keypackage(
-        &self,
-        backend: &MlsCryptoProvider,
-        cs: MlsCiphersuite,
-        ct: MlsCredentialType,
-    ) -> Result<KeyPackage> {
-        let cb = self
-            .find_most_recent_credential_bundle(cs.signature_algorithm(), ct)
-            .await
-            .map_err(RecursiveError::mls_client("finding most recent credential bundle"))?;
-        self.generate_one_keypackage_from_credential_bundle(backend, cs, &cb)
-            .await
-            .map_err(RecursiveError::mls_client(
-                "generating new keypackage from credential bundle",
-            ))
-            .map_err(Into::into)
     }
 }
